+++ tests/kw_config_loader_test.sh
@@ -5,23 +5,40 @@ include './src/kw_config_loader.sh'
 
 KWORKFLOW='kw'
 
-TMP_DIR=tests/.tmp_kw_config_loader_test
+TMP_DIR="$PWD/tests/.tmp_kw_config_loader_test"
 
 function setUp()
 {
   mkdir -p "$TMP_DIR"
-  cp "$PWD/etc/kworkflow_template.config" "$TMP_DIR/kworkflow.config"
+  KW_CONFIG_TEMPLATE="$PWD/etc/init_templates/x86-64/kworkflow_template.config"
+  KW_BUILD_CONFIG_TEMPLATE="$PWD/etc/init_templates/x86-64/build_template.config"
+
+  cp "$KW_CONFIG_TEMPLATE" "$TMP_DIR/kworkflow.config"
+  cp "$KW_BUILD_CONFIG_TEMPLATE" "$TMP_DIR/build.config"
   configurations=()
+  build_config=()
 }
 
 function tearDown()
 {
-  rm -rf "$TMP_DIR"
+  if [[ -d "$TMP_DIR" ]]; then
+    rm -rf "$TMP_DIR"
+  fi
+}
+
+function oneTimeSetUp()
+{
+  mkdir -p "$SHUNIT_TMPDIR"
+}
+
+function oneTimeTearDown()
+{
+  rm -rf "$SHUNIT_TMPDIR" "$TMP_DIR"
 }
 
 function test_parse_configuration_success_exit_code()
 {
-  parse_configuration tests/samples/kworkflow.config
+  parse_configuration "$KW_CONFIG_SAMPLE"
   assertTrue 'kw failed to load a regular config file' "[ 0 -eq $? ]"
 }
 
@@ -30,8 +47,8 @@ function test_parse_configuration_config_with_spaces_and_comments()
   parse_configuration 'tests/samples/kworkflow_space_comments.config'
   assertTrue "($LINENO): Kw failed to load a regular config file" "[ 0 -eq $? ]"
 
-  assertEquals "($LINENO)" "${configurations['arch']}" 'arm64'
-  assertEquals "($LINENO)" "${configurations['kernel_img_name']}" 'Image'
+  assertEquals "($LINENO)" "${configurations['ssh_user']}" 'juca'
+  assertEquals "($LINENO)" "${configurations['mount_point']}" '/home/lala'
   assertEquals "($LINENO)" "${configurations['virtualizer']}" 'libvirt'
   assertEquals "($LINENO)" "${configurations['reboot_after_deploy']}" 'no'
 }
@@ -68,10 +85,8 @@ function assertConfigurations()
 # Test if parse_configuration correctly parses all settings in a file
 function test_parse_configuration_output()
 {
+  # shellcheck disable=2016
   declare -A expected_configurations=(
-    [arch]='arm64'
-    [kernel_img_name]='Image'
-    [cross_compile]='aarch64-linux-gnu-'
     [virtualizer]='libvirt'
     [qemu_path_image]='/home/xpto/p/virty.qcow2'
     [ssh_user]='juca'
@@ -82,7 +97,6 @@ function test_parse_configuration_output()
     [reboot_after_deploy]='no'
     [gui_on]='turn on'
     [gui_off]='turn off'
-    [doc_type]='htmldocs'
     [send_opts]='--annotate --cover-letter --no-chain-reply-to --thread'
     [blocked_emails]='test@email.com'
     [checkpatch_opts]='--no-tree --color=always --strict'
@@ -108,14 +122,10 @@ function test_parse_configuration_output()
   true # Reset return value
 }
 
-# Test if etc/kworkflow_template.config contains all the expected settings
+# Test if etc/init_templates/kworkflow_template.config contains all the expected settings
 function test_parse_configuration_standard_config()
 {
-  # shellcheck disable=2016
   declare -A expected_configurations=(
-    [arch]='x86_64'
-    [kernel_img_name]='bzImage'
-    [menu_config]='nconfig'
     [virtualizer]='qemu-system-x86_64'
     [qemu_path_image]='/home/USERKW/p/virty.qcow2'
     [qemu_hw_options]='-enable-kvm -daemonize -smp 2 -m 1024'
@@ -130,7 +140,6 @@ function test_parse_configuration_standard_config()
     [default_deploy_target]='vm'
     [reboot_after_deploy]='no'
     [disable_statistics_data_track]='no'
-    [doc_type]='htmldocs'
     [send_opts]='--annotate --cover-letter --no-chain-reply-to --thread'
     [checkpatch_opts]='--no-tree --color=always --strict'
     [get_maintainer_opts]='--separator , --nokeywords --nogit --nogit-fallback --norolestats'
@@ -141,7 +150,7 @@ function test_parse_configuration_standard_config()
     [strip_modules_debug_option]='yes'
   )
 
-  parse_configuration "$TMP_DIR/kworkflow.config"
+  parse_configuration "$KW_CONFIG_TEMPLATE"
   assertConfigurations configurations expected_configurations "$LINENO"
 
   true # Reset return value
@@ -175,7 +184,7 @@ function test_parse_configuration_files_loading_order()
   output="$(
     function parse_configuration()
     {
-      printf '%s\n' "$@"
+      printf '%s\n' "$1"
     }
     load_configuration
   )"
@@ -197,7 +206,7 @@ function test_parse_configuration_files_loading_order()
   output="$(
     function parse_configuration()
     {
-      printf '%s\n' "$@"
+      printf '%s\n' "$1"
     }
     load_configuration
   )"
@@ -218,14 +227,17 @@ function test_show_variables_completeness()
 
   # get all assigned options, including commented ones
   # remove #'s and ='s to get option names
-  output="$(cat 'etc/kworkflow_template.config')"
+  output="$(< "$KW_CONFIG_TEMPLATE")"
   output="$(printf '%s\n' "$output" | grep -oE '^(#?\w+=?)' | sed -E 's/[#=]//g')"
 
   for option in $output; do
     possible_options["$option"]='1'
   done
 
-  output="$(show_variables 'TEST_MODE' | grep -E '^    ')"
+  parse_configuration "$KW_CONFIG_TEMPLATE"
+  parse_configuration "$KW_BUILD_CONFIG_TEMPLATE"
+
+  output="$(show_variables 'TEST_MODE' --build | grep -E '^    ')"
   output="$(printf '%s\n' "$output" | sed 's/.*(\(\S*\)).*/\1/')"
 
   for option in $output; do
@@ -234,7 +246,7 @@ function test_show_variables_completeness()
 
   for option in "${!possible_options[@]}"; do
     if [[ ! -v shown_options["$option"] ]]; then
-      fail "($LINENO): show_variables is missing option $option"
+      fail "($LINENO): shown_options is missing option $option"
     fi
   done
 
@@ -252,36 +264,31 @@ function test_show_variables_correctness()
   local value
   local message
 
-  unset configurations
   declare -gA configurations=(
     [ssh_ip]=1
     [ssh_port]=2
     [mount_point]=3
-    [arch]=4
-    [kernel_img_name]=5
-    [cross_compile]=6
-    [menu_config]=7
-    [doc_type]=8
-    [virtualizer]=9
-    [qemu_hw_options]=10
-    [qemu_net_options]=11
-    [qemu_path_image]=12
-    [alert]=13
-    [sound_alert_command]=14
-    [visual_alert_command]=15
-    [default_deploy_target]=16
-    [reboot_after_deploy]=17
-    [disable_statistics_data_track]=18
-    [gui_on]=19
-    [gui_off]=20
+    [menu_config]=4
+    [virtualizer]=5
+    [qemu_hw_options]=6
+    [qemu_net_options]=7
+    [qemu_path_image]=8
+    [alert]=9
+    [sound_alert_command]=10
+    [visual_alert_command]=11
+    [default_deploy_target]=12
+    [reboot_after_deploy]=13
+    [disable_statistics_data_track]=14
+    [gui_on]=15
+    [gui_off]=16
   )
 
-  output="$(show_variables | grep -E '^    ')"
+  output="$(show_variables | grep -E '^\s{3,}')"
 
   while read -r line; do
     option="$(printf '%s\n' "$line" | sed -E 's/.*\((\S*)\).*/\1/')"
     value=$(printf '%s\n' "$line" | sed -E 's/.*: (.*)/\1/')
-    if [[ "${configurations["$option"]}" != "$value" ]]; then
+    if [[ "${configurations[$option]}" != "$value" ]]; then
       message="Value of option $option should be "
       message+="${configurations["$option"]} but is $value"
       fail "($LINENO): $message"
@@ -300,7 +307,7 @@ function test_load_configuration()
     :
   }
 
-  cp "$PWD/etc/kworkflow_template.config" "$SHUNIT_TMPDIR/kworkflow.config"
+  cp "$KW_CONFIG_TEMPLATE" "$SHUNIT_TMPDIR/kworkflow.config"
 
   cd "$SHUNIT_TMPDIR" || {
     fail "($LINENO): It was not possible to move to temporary directory"
@@ -324,7 +331,7 @@ function test_load_configuration()
 
   rm -rf "${SHUNIT_TMPDIR:?}"/*
   mkdir -p "$SHUNIT_TMPDIR/$KW_DIR"
-  cp "$current_path/etc/kworkflow_template.config" "$SHUNIT_TMPDIR/$KW_DIR/kworkflow.config"
+  cp "$KW_CONFIG_TEMPLATE" "$SHUNIT_TMPDIR/$KW_DIR/kworkflow.config"
 
   expected=(
     "1/$CONFIG_FILENAME"
@@ -336,7 +343,7 @@ function test_load_configuration()
   output="$(
     function parse_configuration()
     {
-      printf '%s\n' "$@"
+      printf '%s\n' "$1"
     }
     load_configuration
   )"
